✨ Add support for VANISHED responses
This updates the parser to handle the VANISHED response, and adds a new
response data class: VanishedData.

VanishedData was originally written as a plain class and then converted
to Data.  I'm keeping most of the tests that I wrote prior to converting
it to Data, but I probably wouldn't have written many of them if I had
used Data in the first place.