♻️ Only define DataLite for ruby 3.1
This updates DataLite to more closely match ruby's Data, so it can be a
drop-in replacement (i.e: a polyfill).  On newer ruby versions, it
simply won't be defined at all.  The tests are copied from the
polyfill-data gem and updated they use "Data" as it is resolved inside
the "Net::IMAP" module namespace.  This lets them run with the polyfill
under ruby 3.1 and against the "real" Data in 3.2+.